+++ CHANGES.md
@@ -6,6 +6,7 @@
 * Improved ergonomics of mouse acceleration.
 * `package.bat` now creates Linux-compatible archives.
 * Ctrl and Alt now affect strafe speed and pan sensitivity, respectively. Hold both to greatly reduce strafe speed.
+* Quad views are now somewhat supported. Blender seems to have a slew of its own bugs with them.
 
 # 1.0 (2019-10-01)
 * Initial release