+++ CHANGES.md
@@ -1,4 +1,4 @@
-# 2.0 (2020-12-20)
+# 2.0 (2020-12-22)
 * Pressing C to adjust pivot distance now ignores backfaces when backface culling is enabled in viewport settings.
 * Added camera roll mouse action. Can press R to reset roll.
 * Mouse strafing now remains active while at least one mouse button is still pressed, regardless of whether strafing key was released.
@@ -7,6 +7,7 @@
 * `package.bat` now creates Linux-compatible archives.
 * Ctrl and Alt now affect strafe speed and pan sensitivity, respectively. Hold both to greatly reduce strafe speed.
 * Quad views are now somewhat supported. Blender seems to have a slew of its own bugs with them.
+* The scroll wheel can now be configured to change the field of view.
 
 # 1.0 (2019-10-01)
 * Initial release